pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "met-dsl"
version = "0.1.0"
description = "CLI tooling for the Meteorological DSL IR and Fortran emission pipeline."
authors = [
  { name = "Met DSL Team" }
]
license = { text = "MIT" }
readme = "README.md"
dependencies = [
  "typer>=0.4.0",
  "pydantic>=1.9.0",
  "jinja2>=3.0.0",
  "rich>=10.16.0",
  "pyyaml>=5.4.1"
]
requires-python = ">=3.9"

[project.optional-dependencies]
cli = [
  "click>=8.1.0"
]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.0",
  "black>=24.4.0",
  "mypy>=1.10.0"
]

[project.scripts]
metdsl = "metdsl.cli.emit:app"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
target-version = "py39"
extend-select = ["I", "N"]

[tool.pytest.ini_options]
addopts = "--maxfail=1 --disable-warnings"
pythonpath = ["src"]

[tool.coverage.run]
branch = true
source = ["metdsl"]

[tool.hatch.build.targets.wheel]
packages = ["src/metdsl"]

[tool.hatch.build.targets.sdist]
include = [
  "src/metdsl",
  "tests",
  "docs",
  "scripts",
  "pyproject.toml",
  "README.md",
]
